cybersecurity_assessment_tool/requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.11.0
arrow==1.4.0
asgiref==3.9.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
blessed==1.33.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
cryptography==46.0.5
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1764921340/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dj-database-url==3.1.2
Django==5.2.6
django-debug-toolbar==6.2.0
django-extensions==4.1
django-fernet-encrypted-fields==0.3.1
django-picklefield==3.4.0
django-q==1.3.9
django-q2==1.9.0
django-recurrence==1.11.1
djangorestframework==3.16.1
dnspython==2.7.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
Faker==40.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.39.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==25.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jinxed==1.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver==1.5.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
proto-plus==1.26.1
protobuf==5.29.5
psutil @ file:///D:/bld/psutil_1758169180033/work
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
tenacity==9.1.2
tornado @ file:///D:/bld/tornado_1762506734045/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.6.0
websockets==15.0.1
whitenoise==6.12.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
